+++ relation_engine_server/api_versions/api_v1.py
@@ -7,6 +7,7 @@ from relation_engine_server.utils import (
     pull_spec,
     config,
     parse_json,
+    ensure_specs,
 )
 from relation_engine_server.utils.json_validation import run_validator
 from relation_engine_server.exceptions import InvalidParameters
@@ -193,6 +194,30 @@ def show_config():
             "spec_release_path": conf["spec_release_path"],
         }
     )
+
+
+@api_v1.route("/ensure_specs", methods=["GET"])
+def ensure_all_specs():
+    """
+    Ensure that the local index/view/analyzer specs under spec/ have a
+    corresponding spec on the server.
+
+    This endpoint is not strictly necessary, as the ensure_specs.ensure_all()
+    code should triggered in startup scripts. This is more insurance in case
+    one wishes to ensure the specs without re-deployment
+
+    Example ensure_specs.ensure_all() return value:
+    {
+        "indexes": [],
+        "views": ["Compounds/arangosearch", "Reactions/arangosearch"],
+        "analyzers": ["icu_tokenize/text"]
+    }
+    """
+    failed_names = ensure_specs.ensure_all()
+    if any([name for schema_type, names in failed_names.items() for name in names]):
+        return flask.jsonify(failed_names), 500
+    else:
+        return flask.jsonify(failed_names)
 
 
 def _preprocess_stored_query(query_text, config):