+++ relation_engine_server/utils/pull_spec.py
@@ -4,12 +4,13 @@ import tarfile
 import tempfile
 import shutil
 import json
-import glob
 import yaml
 from typing import Optional
 
 from relation_engine_server.utils import arango_client
 from relation_engine_server.utils.config import get_config
+from relation_engine_server.utils.ensure_specs import ensure_all
+from spec.validate import get_schema_type_paths
 
 _CONF = get_config()
 
@@ -59,13 +60,21 @@ def download_specs(
         do_init_collections()
         do_init_views()
         do_init_analyzers()
+    # Check that local specs have matching server specs
+    # Necessary because creating resources like indexes
+    # does not overwrite any pre-existing indexes
+    failed_names = ensure_all()
+    if any([name for schema_type, names in failed_names.items() for name in names]):
+        raise RuntimeError(
+            "Some local specs have no matching server specs:"
+            "\n" + json.dumps(failed_names, indent=4)
+        )
     return update_name
 
 
 def do_init_collections():
     """Initialize any uninitialized collections in the database from a set of collection schemas."""
-    pattern = os.path.join(_CONF["spec_paths"]["collections"], "**", "*.yaml")
-    for path in glob.iglob(pattern):
+    for path in get_schema_type_paths("collection"):
         coll_name = os.path.basename(os.path.splitext(path)[0])
         with open(path) as fd:
             config = yaml.safe_load(fd)
@@ -74,8 +83,7 @@ def do_init_collections():
 
 def do_init_views():
     """Initialize any uninitialized views in the database from a set of schemas."""
-    pattern = os.path.join(_CONF["spec_paths"]["views"], "**", "*.json")
-    for path in glob.iglob(pattern):
+    for path in get_schema_type_paths("view"):
         view_name = os.path.basename(os.path.splitext(path)[0])
         with open(path) as fd:
             config = json.load(fd)
@@ -83,8 +91,7 @@ def do_init_views():
 
 
 def do_init_analyzers():
-    pattern = os.path.join(_CONF["spec_paths"]["analyzers"], "*.json")
-    for path in glob.iglob(pattern):
+    for path in get_schema_type_paths("analyzer"):
         analyzer_name = os.path.basename(os.path.splitext(path)[0])
         with open(path) as fd:
             config = json.load(fd)