+++ CHANGELOG.md
@@ -6,11 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+
+## [0.0.17] - 2022-01-25
+### Added
+- Ensure local specs match server specs
+### Changed
+- Remove explicit namespace from "icu_tokenize" analyzer
+
 ## [0.0.16] - 2022-01-14
 ### Added
 - Added github actions to build docker images on ghcr.io
-
-
+- Added generic fulltext search stored query
 
 ## [0.0.12] - 2021-01-29
 ### Added